requirements.txt
docker==6.0.1
requests==2.29.0
tqdm==4.65.0
python-dotenv
